+++ spec/02-integration/05-proxy/02-router_spec.lua
@@ -159,7 +159,7 @@ for _, strategy in helpers.each_strategy() do
         plugins = "bundled,enable-buffering",
         nginx_conf = "spec/fixtures/custom_nginx.template",
         stream_listen = string.format("127.0.0.1:%d ssl", stream_tls_listen_port),
-        debug_header = true,
+        allow_debug_header = true,
       }, nil, nil, fixtures))
     end)
 
@@ -2279,7 +2279,7 @@ for _, strategy in helpers.each_strategy() do
           nginx_worker_processes = 4,
           plugins = "bundled,enable-buffering",
           nginx_conf = "spec/fixtures/custom_nginx.template",
-          debug_header = true,
+          allow_debug_header = true,
         }))
       end)
 
@@ -2390,7 +2390,7 @@ for _, strategy in helpers.each_strategy() do
     end)
   end
 
-  describe("disable debug_header config" , function()
+  describe("disable allow_debug_header config" , function()
     local proxy_client
 
     lazy_setup(function()
@@ -2438,7 +2438,7 @@ for _, strategy in helpers.each_strategy() do
       end
     end)
 
-    it("disable debug_header config", function()
+    it("disable allow_debug_header config", function()
       for _ = 1, 1000 do
         proxy_client = helpers.proxy_client()
         local res = assert(proxy_client:send {